Start cli for wsdot report
